Module skrifa::attribute

Primary attributes typically used for font classification and selection.

Structs

Attributes

Stretch, style and weight attributes of a font.

Stretch

Visual width of a font– a relative change from the normal aspect ratio, typically in the range 0.5 to 2.0.

Weight

Visual weight class of a font, typically on a scale from 1.0 to 1000.0.

Enums

Style

Visual style or ‘slope’ of a font.