skrifa

Module bitmap

Bitmap strikes and glyphs.

Structs

BitmapGlyph

An embedded bitmap glyph.

BitmapStrike

Set of embedded bitmap glyphs of a specific size.

BitmapStrikes

Set of strikes, each containing embedded bitmaps of a single size.

MaskData

A single channel alpha mask.

Enums

BitmapData

Data content of a bitmap.

BitmapFormat

The format (or table) containing the data backing a set of bitmap strikes.

Origin

The origin point for drawing a bitmap glyph.