skrifa::color

Module instance

COLR table instance.

Structs

ColorStops

Collection of ColorStop or VarColorStop.

ColrInstance

Combination of a COLR table and a location in variation space for resolving paints.

ResolvedColorStop

Simplified version of a ColorStop or VarColorStop with applied variation deltas.

Enums

ResolvedPaint

Simplified version of Paint with applied variation deltas.

Functions

resolve_clip_box

Resolves a clip box, applying variation deltas using the given instance.

resolve_paint

Resolves this paint with the given instance.

Type Aliases

PaintId

Unique paint identifier used for detecting cycles in the paint graph.