skrifa::color::instance

Function resolve_paint

pub fn resolve_paint<'a>(
    instance: &ColrInstance<'a>,
    paint: &Paint<'a>,
) -> Result<ResolvedPaint<'a>, ReadError>

Resolves this paint with the given instance.

Resolving means that all numeric values are converted to 32-bit floating point, variation deltas are applied (also computed fully in floating point), and the various transform paints are collapsed into a single value for their category (transform, translate, scale, rotate and skew).

This provides a simpler type for consumers that are more interested in extracting the semantics of the graph rather than working with the raw encoded structures.