skrifa::color

Module traversal

Structs

TraversalState 🔒

Constants

MAX_INLINE_COLOR_STOPS 🔒

MAX_NODES 🔒

Maximum number of nodes visited during a single traversal.

MAX_TRAVERSAL_DEPTH 🔒

Depth at which we will stop traversing and return an error.

Functions

get_clipbox_font_units 🔒

traverse_unresolved_paint 🔒

traverse_v0_range 🔒

traverse_with_callbacks 🔒

Type Aliases

ColorStopVec 🔒

PaintDecycler 🔒