skrifa::metrics
Type Alias BoundingBox

pub type BoundingBox = BoundingBox<f32>;

Type for a bounding box with single precision floating point coordinates.

Aliased Type

struct BoundingBox {
    pub x_min: f32,
    pub y_min: f32,
    pub x_max: f32,
    pub y_max: f32,
}

Fields
x_min: f32

Minimum extent in the x direction– the left side of a region.

y_min: f32

Minimum extent in the y direction. In a Y-up coordinate system, which is used by fonts, this represents the bottom of a region.

x_max: f32

Maximum extent in the x direction– the right side of a region.

y_max: f32

Maximum extend in the y direction. In a Y-up coordinate system, which is used by fonts, this represents the top of the region.