skrifa::outline

Module autohint

Runtime autohinting support.

Structs

Axis

Segments and edges for one dimension of an outline.

BlueProvenance

Source for an alignment zone.

BlueZones

Defines the zone(s) that are associated with a blue value.

Edge

Sequence of segments used for grid-fitting.

EdgeHint

Hinting information for an edge.

GlyphStyle

Defines the script and style associated with a single glyph.

GlyphStyles

Set of derived glyph styles that are used for automatic hinting.

HintPlan

Plan for hinting an outline.

PointHint

Hinting information for a point.

Scale

Captures scaling parameters which may be modified during metrics computation.

ScaleFlags

Flags that define how scaling and hinting is applied.

ScaledAxisMetrics

Scaled metrics for a single axis.

ScaledBlue

A scaled alignment zone.

ScaledStyleMetrics

Scaled metrics for a single style and script.

ScaledWidth

A scaled stem width.

ScriptClass

Defines the basic properties for each script supported by the autohinter.

Segment

Sequence of points with a single dominant direction.

StyleClass

Defines the basic properties for each style supported by the autohinter.

TopoFlags

Flags that define the properties of segments and edges.

UnscaledAxisMetrics

Unscaled metrics for a single axis.

UnscaledBlue

An unscaled alignment zone.

UnscaledStyleMetrics

Unscaled metrics for a single style and script.

WidthMetrics

Metrics for the set of stems along a single axis.

Enums

Dimension

The dimension of an axis.

Direction

Hinting directions.

EdgeAction

Hinting action for an edge.

HintAction

Hinting action for a point or an edge.

PointAction

Hinting action for a point.

ScriptGroup

Determines which algorithms the autohinter will use while generating metrics and processing a glyph outline.

Constants

SCRIPT_CLASSES

STYLE_CLASSES