skrifa::outline::autohint::hint

Module edges

Edge hinting.

Let’s actually do some grid fitting. Here we align edges to the pixel grid. This is the final step before applying the edge adjustments to the original outline points.

Enums

LinkDir 🔒

Functions

adjust_link 🔒

Helper to adjust links based on hinting direction.

align_edges_to_blues 🔒

Align horizontal edges to blue zones.

align_linked_edge 🔒

Align one stem edge relative to previous stem edge.

align_remaining_edges 🔒

Align serif and single segment edges.

align_serif_edge 🔒

Shift the serif edge by the adjustment made to base edge.

align_stem_edges 🔒

Align stem edges, trying to main relative order of stems in the glyph.

find_bounding_completed_edges 🔒

Returns the indices of the “completed” edges before and after the given edge index.

hint_edges 🔒

Main Latin grid-fitting routine.

hint_lowercase_m 🔒

Make sure that lowercase m’s maintain symmetry.

hint_normal_stem_cjk 🔒

Adjusts both edges of a stem and returns the delta.

latin_remaining_bounds 🔒

snap_width 🔒

Snap a scaled width to one of the standard widths.

stem_width 🔒

Compute the snapped width of a given stem.