skrifa::outline::autohint::hint

Module outline

Apply edge hints to an outline.

This happens in three passes:

Align points that are directly attached to edges. These are the points which originally generated the edge and are coincident with the edge coordinate (within a threshold) for a given axis. This may include points that were originally classified as weak.

Interpolate non-weak points that were not touched by the previous pass. This searches for the edges that enclose the point and interpolates the coordinate based on the adjustment applied to those edges.

Interpolate remaining untouched points. These are generally the weak points: those that are very near other points or lacking a dominant inward or outward direction.

The final result is a fully hinted outline.

Functions

align_edge_points 🔒

Align all points of an edge to the same coordinate value.

align_strong_points 🔒

Align the strong points; equivalent to the TrueType IP instruction.

align_weak_points 🔒

Align the weak points; equivalent to the TrueType IUP instruction.

iup_interpolate 🔒

Interpolate the original coordinates all of points between p1_ix and p2_ix (inclusive) to get hinted coordinates, using the points at ref1_ix and ref2_ix as the reference points.

iup_shift 🔒

Shift original coordinates of all points between p1_ix and p2_ix (inclusive) to get hinted coordinates using the same difference as given by the point at ref_ix.

store_point 🔒