skrifa::outline::autohint::metrics

Module blues

Latin blue values.

Structs

BlueZones

Defines the zone(s) that are associated with a blue value.

ScaledBlue

A scaled alignment zone.

UnscaledBlue

An unscaled alignment zone.

Constants

BLUE_STRING_MAX_LEN 🔒

MAX_BLUES 🔒

Maximum number of blue values.

MAX_INLINE_POINTS 🔒

Functions

buffers 🔒

compute_cjk_blues 🔒

Compute unscaled blue values for the CJK script set.

compute_default_blues 🔒

Compute unscaled blue values for the default script set.

compute_unscaled_blues 🔒

Compute unscaled blues values for each axis.

cycle_backward 🔒

Iterator that begins at start - 1 and cycles through all items of the slice in reverse order, ending with start.

cycle_forward 🔒

Iterator that begins at start + 1 and cycles through all items of the slice in forward order, ending with start.

satisfies_min_long_segment_len 🔒

Given inclusive indices and a contour length, returns true if the segment is of sufficient size to test for bumps when detecting “long” Hebrew alignment zones.

things_all_blues_need 🔒

A thneed is something everyone needs

Type Aliases

ScaledBlues 🔒

UnscaledBlues 🔒