skrifa::outline::autohint::metrics::widths

Function compute_widths

pub(super) fn compute_widths(
    shaper: &Shaper<'_>,
    coords: &[F2Dot14],
    style: &StyleClass,
    quirks: QuirksMode,
) -> [(WidthMetrics, SmallVec<i32, MAX_WIDTHS>); 2]

Compute all stem widths and initialize standard width and height for the given script.

Returns width metrics and unscaled widths for each dimension.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/57617782464411201ce7bbc93b086c1b4d7d84a5/src/autofit/aflatin.c#L54