skrifa::outline::autohint

Struct Segment

pub struct Segment {

    pub(crate) flags: TopoFlags,
    pub(crate) dir: Direction,
    pub(crate) pos: i16,
    pub(crate) delta: i16,
    pub(crate) min_coord: i16,
    pub(crate) max_coord: i16,
    pub(crate) height: i16,
    pub(crate) score: i32,
    pub(crate) len: i32,
    pub(crate) link_ix: Option<u16>,
    pub(crate) serif_ix: Option<u16>,
    pub(crate) first_ix: u16,
    pub(crate) last_ix: u16,
    pub(crate) edge_ix: Option<u16>,
    pub(crate) edge_next_ix: Option<u16>,

}

Sequence of points with a single dominant direction.

Fields
flags: TopoFlags

Flags describing the properties of the segment.

dir: Direction

Dominant direction of the segment.

pos: i16

Position of the segment.

delta: i16

Deviation from segment position.

min_coord: i16

Minimum coordinate of the segment.

max_coord: i16

Maximum coordinate of the segment.

height: i16

Hinted segment height.

score: i32

Used during stem matching.

len: i32

Used during stem matching.

link_ix: Option<u16>

Index of best candidate for a stem link.

serif_ix: Option<u16>

Index of best candidate for a serif link.

first_ix: u16

Index of first point in the outline.

last_ix: u16

Index of last point in the outline.

edge_ix: Option<u16>

Index of edge that is associated with the segment.

edge_next_ix: Option<u16>

Index of next segment in edge’s segment list.

Implementations

impl Segment

pub fn flags(&self) -> TopoFlags

Returns flags describing the properties of the segment.

pub fn direction(&self) -> Direction

Returns the dominant direction of the segment.

pub fn position(&self) -> i16

Returns the computed position of the segment.

pub fn delta(&self) -> i16

Returns the deviation from the computed position.

pub fn min_coord(&self) -> i16

Returns the minimum coordinate of the segment.

pub fn max_coord(&self) -> i16

Returns the maximum coordinate of the segment.

pub fn height(&self) -> i16

Returns the hinted height of the segment.

pub fn score(&self) -> i32

Returns the computed score of the segment; used during stem matching.

pub fn length(&self) -> i32

Returns the computed length of the segment; used during stem matching.

pub fn link_index(&self) -> Option<u16>

Returns the index of the best candidate for a stem link.

pub fn serif_index(&self) -> Option<u16>

Returns the index of the best candidate for a serif link.

pub fn point_indices(&self) -> (u16, u16)

Returns the indices of first and last points that define the segment.

pub fn edge_index(&self) -> Option<u16>

Returns the index of edge that is associated with the segment.

pub fn next_in_edge_index(&self) -> Option<u16>

Returns the index of next segment in the associated edge’s segment list.

impl Segment

pub(crate) fn first(&self) -> usize

pub(crate) fn first_point<'a>(&self, points: &'a [Point]) -> &'a Point

pub(crate) fn last(&self) -> usize

pub(crate) fn last_point<'a>(&self, points: &'a [Point]) -> &'a Point

pub(crate) fn edge<'a>(&self, edges: &'a [Edge]) -> Option<&'a Edge>

pub(crate) fn next_in_edge<'a>( &self, segments: &'a [Segment], ) -> Option<&'a Segment>

Returns the next segment in this segment’s parent edge.

pub(crate) fn link<'a>(&self, segments: &'a [Segment]) -> Option<&'a Segment>

pub(crate) fn adjust_height(&mut self, v1: i32, v2: i32)

Adjust the height of the segment by half the distance between v1 and v2 using wrapping arithmetic.