skrifa::outline::autohint

Struct StyleClass

pub struct StyleClass {
    pub name: &'static str,
    pub index: usize,
    pub script: &'static ScriptClass,
    pub feature: Option<Tag>,
}

Defines the basic properties for each style supported by the autohinter.

There’s mostly a 1:1 correspondence between styles and scripts except in the cases where style coverage is determined by OpenType feature coverage.

Fields
name: &'static str

The name of the style class.

index: usize

Index of self in the STYLE_CLASSES array.

script: &'static ScriptClass

Associated Unicode script.

feature: Option<Tag>

OpenType feature tag for styles that derive coverage from layout tables.

Implementations

impl StyleClass

pub const ADLM: usize = 0

pub const ARAB: usize = 1

pub const ARMN: usize = 2

pub const AVST: usize = 3

pub const BAMU: usize = 4

pub const BENG: usize = 5

pub const BUHD: usize = 6

pub const CAKM: usize = 7

pub const CANS: usize = 8

pub const CARI: usize = 9

pub const CHER: usize = 10

pub const COPT: usize = 11

pub const CPRT: usize = 12

pub const CYRL_C2CP: usize = 13

pub const CYRL_C2SC: usize = 14

pub const CYRL_ORDN: usize = 15

pub const CYRL_PCAP: usize = 16

pub const CYRL_RUBY: usize = 17

pub const CYRL_SINF: usize = 18

pub const CYRL_SMCP: usize = 19

pub const CYRL_SUBS: usize = 20

pub const CYRL_SUPS: usize = 21

pub const CYRL_TITL: usize = 22

pub const CYRL: usize = 23

pub const DEVA: usize = 24

pub const DSRT: usize = 25

pub const ETHI: usize = 26

pub const GEOR: usize = 27

pub const GEOK: usize = 28

pub const GLAG: usize = 29

pub const GOTH: usize = 30

pub const GREK_C2CP: usize = 31

pub const GREK_C2SC: usize = 32

pub const GREK_ORDN: usize = 33

pub const GREK_PCAP: usize = 34

pub const GREK_RUBY: usize = 35

pub const GREK_SINF: usize = 36

pub const GREK_SMCP: usize = 37

pub const GREK_SUBS: usize = 38

pub const GREK_SUPS: usize = 39

pub const GREK_TITL: usize = 40

pub const GREK: usize = 41

pub const GUJR: usize = 42

pub const GURU: usize = 43

pub const HEBR: usize = 44

pub const HMNP: usize = 45

pub const KALI: usize = 46

pub const KHMR: usize = 47

pub const KHMS: usize = 48

pub const KNDA: usize = 49

pub const LAOO: usize = 50

pub const LATN_C2CP: usize = 51

pub const LATN_C2SC: usize = 52

pub const LATN_ORDN: usize = 53

pub const LATN_PCAP: usize = 54

pub const LATN_RUBY: usize = 55

pub const LATN_SINF: usize = 56

pub const LATN_SMCP: usize = 57

pub const LATN_SUBS: usize = 58

pub const LATN_SUPS: usize = 59

pub const LATN_TITL: usize = 60

pub const LATN: usize = 61

pub const LATB: usize = 62

pub const LATP: usize = 63

pub const LISU: usize = 64

pub const MLYM: usize = 65

pub const MEDF: usize = 66

pub const MONG: usize = 67

pub const MYMR: usize = 68

pub const NKOO: usize = 69

pub const NONE: usize = 70

pub const OLCK: usize = 71

pub const ORKH: usize = 72

pub const OSGE: usize = 73

pub const OSMA: usize = 74

pub const ROHG: usize = 75

pub const SAUR: usize = 76

pub const SHAW: usize = 77

pub const SINH: usize = 78

pub const SUND: usize = 79

pub const TAML: usize = 80

pub const TAVT: usize = 81

pub const TELU: usize = 82

pub const TFNG: usize = 83

pub const THAI: usize = 84

pub const VAII: usize = 85

pub const LIMB: usize = 86

pub const ORYA: usize = 87

pub const SYLO: usize = 88

pub const TIBT: usize = 89

pub const HANI: usize = 90