skrifa::outline::autohint

Module style

Styles, scripts and glyph style mapping.

Structs

GlyphStyle

Defines the script and style associated with a single glyph.

GlyphStyleMap 🔒

Maps glyph identifiers to glyph styles.

ScriptClass

Defines the basic properties for each script supported by the autohinter.

StyleClass

Defines the basic properties for each style supported by the autohinter.

StyleRange 🔒

Associates a basic glyph style with a range of codepoints.

Enums

ScriptGroup

Determines which algorithms the autohinter will use while generating metrics and processing a glyph outline.

Constants

MAX_STYLES 🔒

SCRIPT_CLASSES

STYLE_CLASSES

STYLE_RANGES 🔒

UNMAPPED_STYLE 🔒

Sentinel for unused styles in GlyphStyleMap::metrics_map.

Functions

base_range 🔒

non_base_range 🔒