skrifa::outline::autohint::style

Struct StyleClass

pub(crate) struct StyleClass {
    pub name: &'static str,
    pub index: usize,
    pub script: &'static ScriptClass,
    pub feature: Option<Tag>,
}

Defines the basic properties for each style supported by the autohinter.

There’s mostly a 1:1 correspondence between styles and scripts except in the cases where style coverage is determined by OpenType feature coverage.

Fields
name: &'static strindex: usize

Index of self in the STYLE_CLASSES array.

script: &'static ScriptClass

Associated Unicode script.

feature: Option<Tag>

OpenType feature tag for styles that derive coverage from layout tables.

Implementations

impl StyleClass

pub(crate) fn from_index(index: u16) -> Option<&'static StyleClass>

impl StyleClass

pub const ADLM: usize = 0usize

pub const ARAB: usize = 1usize

pub const ARMN: usize = 2usize

pub const AVST: usize = 3usize

pub const BAMU: usize = 4usize

pub const BENG: usize = 5usize

pub const BUHD: usize = 6usize

pub const CAKM: usize = 7usize

pub const CANS: usize = 8usize

pub const CARI: usize = 9usize

pub const CHER: usize = 10usize

pub const COPT: usize = 11usize

pub const CPRT: usize = 12usize

pub const CYRL_C2CP: usize = 13usize

pub const CYRL_C2SC: usize = 14usize

pub const CYRL_ORDN: usize = 15usize

pub const CYRL_PCAP: usize = 16usize

pub const CYRL_RUBY: usize = 17usize

pub const CYRL_SINF: usize = 18usize

pub const CYRL_SMCP: usize = 19usize

pub const CYRL_SUBS: usize = 20usize

pub const CYRL_SUPS: usize = 21usize

pub const CYRL_TITL: usize = 22usize

pub const CYRL: usize = 23usize

pub const DEVA: usize = 24usize

pub const DSRT: usize = 25usize

pub const ETHI: usize = 26usize

pub const GEOR: usize = 27usize

pub const GEOK: usize = 28usize

pub const GLAG: usize = 29usize

pub const GOTH: usize = 30usize

pub const GREK_C2CP: usize = 31usize

pub const GREK_C2SC: usize = 32usize

pub const GREK_ORDN: usize = 33usize

pub const GREK_PCAP: usize = 34usize

pub const GREK_RUBY: usize = 35usize

pub const GREK_SINF: usize = 36usize

pub const GREK_SMCP: usize = 37usize

pub const GREK_SUBS: usize = 38usize

pub const GREK_SUPS: usize = 39usize

pub const GREK_TITL: usize = 40usize

pub const GREK: usize = 41usize

pub const GUJR: usize = 42usize

pub const GURU: usize = 43usize

pub const HEBR: usize = 44usize

pub const KALI: usize = 45usize

pub const KHMR: usize = 46usize

pub const KHMS: usize = 47usize

pub const KNDA: usize = 48usize

pub const LAOO: usize = 49usize

pub const LATN_C2CP: usize = 50usize

pub const LATN_C2SC: usize = 51usize

pub const LATN_ORDN: usize = 52usize

pub const LATN_PCAP: usize = 53usize

pub const LATN_RUBY: usize = 54usize

pub const LATN_SINF: usize = 55usize

pub const LATN_SMCP: usize = 56usize

pub const LATN_SUBS: usize = 57usize

pub const LATN_SUPS: usize = 58usize

pub const LATN_TITL: usize = 59usize

pub const LATN: usize = 60usize

pub const LATB: usize = 61usize

pub const LATP: usize = 62usize

pub const LISU: usize = 63usize

pub const MLYM: usize = 64usize

pub const MEDF: usize = 65usize

pub const MONG: usize = 66usize

pub const MYMR: usize = 67usize

pub const NKOO: usize = 68usize

pub const NONE: usize = 69usize

pub const OLCK: usize = 70usize

pub const ORKH: usize = 71usize

pub const OSGE: usize = 72usize

pub const OSMA: usize = 73usize

pub const ROHG: usize = 74usize

pub const SAUR: usize = 75usize

pub const SHAW: usize = 76usize

pub const SINH: usize = 77usize

pub const SUND: usize = 78usize

pub const TAML: usize = 79usize

pub const TAVT: usize = 80usize

pub const TELU: usize = 81usize

pub const TFNG: usize = 82usize

pub const THAI: usize = 83usize

pub const VAII: usize = 84usize

pub const LIMB: usize = 85usize

pub const ORYA: usize = 86usize

pub const SYLO: usize = 87usize

pub const TIBT: usize = 88usize

pub const HANI: usize = 89usize