skrifa::outline

Module cff

Support for scaling CFF outlines.

Modules

hint 🔒

CFF hinting.

Structs

CharstringEvaluator 🔒

FontDict 🔒

Entries that we parse from a Font DICT.

HintedTransformingSink 🔒

Command sink adapter that applies a transform to hinted coordinates.

NopFilteringSink 🔒

Command sink adapter that suppresses degenerate move and line commands.

Outlines 🔒

Type for loading, scaling and hinting outlines in CFF/CFF2 tables.

PenSink 🔒

Command sink that sends the results of charstring evaluation to an OutlinePen.

PrivateDict 🔒

Entries that we parse from the Private DICT to support charstring evaluation.

ScalingSink26Dot6 🔒

Command sink adapter that applies a scaling factor.

ScalingTransformingSink 🔒

Command sink adapter that applies both a transform and a scaling factor.

Subfont 🔒

Specifies local subroutines and hinting parameters for some subset of glyphs in a CFF or CFF2 table.

TopDict 🔒

Entries that we parse from the Top DICT that are required to support charstring evaluation.

Enums

PendingElement 🔒

Constants

ONE_OVER_64 🔒

Functions

matrix_mul_scaled 🔒

Simple fixed point matrix multiplication with a scaling factor.

transform 🔒