skrifa::outline::glyf::hint

Module engine

TrueType bytecode interpreter.

Modules

arith 🔒

Arithmetic and math instructions.

control_flow 🔒

Managing the flow of control.

cvt 🔒

Managing the control value table.

data 🔒

Reading and writing data.

definition 🔒

Defining and using functions and instructions.

delta 🔒

Managing delta exceptions.

dispatch 🔒

Instruction decoding and dispatch.

graphics 🔒

Managing the graphics state.

logical 🔒

Logical functions.

misc 🔒

Miscellaneous instructions.

outline 🔒

Managing outlines.

round 🔒

Compensating for the engine characteristics (rounding).

stack 🔒

Managing the stack and pushing data onto the interpreter stack.

storage 🔒

Managing the storage area.

Structs

Engine

TrueType bytecode interpreter.

WorkBudget 🔒

Tracks budgets for control flow to limit execution time.

Constants

MAX_RUN_INSTRUCTIONS 🔒

Maximum number of instructions we will execute in Engine::run(). This is used to ensure termination of a hinting program. See https://gitlab.freedesktop.org/freetype/freetype/-/blob/57617782464411201ce7bbc93b086c1b4d7d84a5/include/freetype/config/ftoption.h#L744

Type Aliases

OpResult