skrifa::outline::glyf::hint::engine

Struct WorkBudget

struct WorkBudget {
    loop_limit: usize,
    backward_jumps: usize,
    loop_calls: usize,
    skipped: usize,
}

Tracks budgets for control flow to limit execution time.

Fields
loop_limit: usize

Maximum number of times we can do backward jumps or loop calls.

backward_jumps: usize

Current number of backward jumps executed.

loop_calls: usize

Current number of loop call iterations executed.

skipped: usize

Counts number of instructions skipped when handling if/else conditions.

Implementations

impl WorkBudget

fn new(outlines: &Outlines<'_>, point_count: Option<usize>) -> Self

fn reset(&mut self)

fn doing_backward_jump(&mut self) -> Result<(), HintErrorKind>

fn doing_loop_call(&mut self, count: usize) -> Result<(), HintErrorKind>

fn skipping_instruction(&mut self) -> Result<(), HintErrorKind>