skrifa::outline::glyf::hint::instance

Struct HintInstance

pub struct HintInstance {
    functions: Vec<Definition>,
    instructions: Vec<Definition>,
    cvt: Vec<i32>,
    storage: Vec<i32>,
    graphics: RetainedGraphicsState,
    twilight_scaled: Vec<Point<F26Dot6>>,
    twilight_original_scaled: Vec<Point<F26Dot6>>,
    twilight_flags: Vec<PointFlags>,
    axis_count: u16,
    max_stack: usize,
}

Fields
functions: Vec<Definition>instructions: Vec<Definition>cvt: Vec<i32>storage: Vec<i32>graphics: RetainedGraphicsStatetwilight_scaled: Vec<Point<F26Dot6>>twilight_original_scaled: Vec<Point<F26Dot6>>twilight_flags: Vec<PointFlags>axis_count: u16max_stack: usize
Implementations

impl HintInstance

pub fn reconfigure( &mut self, outlines: &Outlines<'_>, scale: i32, ppem: i32, target: Target, coords: &[F2Dot14], ) -> Result<(), HintError>

pub fn is_enabled(&self) -> bool

Returns true if we should actually apply hinting.

Hinting can be completely disabled by the control value program.

pub fn backward_compatibility(&self) -> bool

Returns true if backward compatibility mode has been activated by the hinter settings or the prep table.

pub fn hint( &self, outlines: &Outlines<'_>, outline: &mut HintOutline<'_>, is_pedantic: bool, ) -> Result<(), HintError>

fn setup(&mut self, outlines: &Outlines<'_>, scale: i32, coords: &[F2Dot14])

Captures limits, resizes buffers and scales the CVT.