skrifa::outline::glyf

Struct FreeTypeScaler

pub(crate) struct FreeTypeScaler<'a> {

    outlines: &'a Outlines<'a>,
    memory: FreeTypeOutlineMemory<'a>,
    coords: &'a [F2Dot14],
    point_count: usize,
    contour_count: usize,
    component_delta_count: usize,
    ppem: f32,
    scale: F26Dot6,
    is_scaled: bool,
    is_hinted: bool,
    pedantic_hinting: bool,
    phantom: [Point<F26Dot6>; 4],
    hinter: Option<&'a HintInstance>,

}

F26Dot6 coords, Fixed deltas, and a penchant for rounding

Fields
outlines: &'a Outlines<'a>memory: FreeTypeOutlineMemory<'a>coords: &'a [F2Dot14]point_count: usizecontour_count: usizecomponent_delta_count: usizeppem: f32scale: F26Dot6is_scaled: boolis_hinted: boolpedantic_hinting: boolphantom: [Point<F26Dot6>; 4]

Phantom points. These are 4 extra points appended to the end of an outline that allow the bytecode interpreter to produce hinted metrics.

See https://learn.microsoft.com/en-us/typography/opentype/spec/tt_instructing_glyphs#phantom-points

hinter: Option<&'a HintInstance>
Implementations

impl<'a> FreeTypeScaler<'a>

pub(crate) fn unhinted( outlines: &'a Outlines<'a>, outline: &'a Outline<'_>, buf: &'a mut [u8], ppem: Option<f32>, coords: &'a [F2Dot14], ) -> Result<Self, DrawError>

pub(crate) fn hinted( outlines: &'a Outlines<'a>, outline: &'a Outline<'_>, buf: &'a mut [u8], ppem: Option<f32>, coords: &'a [F2Dot14], hinter: &'a HintInstance, pedantic_hinting: bool, ) -> Result<Self, DrawError>

pub(crate) fn scale( self, glyph: &Option<Glyph<'_>>, glyph_id: GlyphId, ) -> Result<ScaledOutline<'a, F26Dot6>, DrawError>