skrifa::outline::glyf

Struct HarfBuzzScaler

pub(crate) struct HarfBuzzScaler<'a> {
    outlines: &'a Outlines<'a>,
    memory: HarfBuzzOutlineMemory<'a>,
    coords: &'a [F2Dot14],
    point_count: usize,
    contour_count: usize,
    component_delta_count: usize,
    ppem: f32,
    scale: f32,
    phantom: [Point<f32>; 4],
}

f32 all the things. Hold your rounding. No hinting.

Fields
outlines: &'a Outlines<'a>memory: HarfBuzzOutlineMemory<'a>coords: &'a [F2Dot14]point_count: usizecontour_count: usizecomponent_delta_count: usizeppem: f32scale: f32phantom: [Point<f32>; 4]

Phantom points. These are 4 extra points appended to the end of an outline that allow the bytecode interpreter to produce hinted metrics.

See https://learn.microsoft.com/en-us/typography/opentype/spec/tt_instructing_glyphs#phantom-points

Implementations

impl<'a> HarfBuzzScaler<'a>

pub(crate) fn unhinted( outlines: &'a Outlines<'a>, outline: &'a Outline<'_>, buf: &'a mut [u8], ppem: Option<f32>, coords: &'a [F2Dot14], ) -> Result<Self, DrawError>

pub(crate) fn scale( self, glyph: &Option<Glyph<'_>>, glyph_id: GlyphId, ) -> Result<ScaledOutline<'a, f32>, DrawError>