skrifa::outline

Module hint_reliant

Name detection for fonts that require hinting to be run for correct contours (FreeType calls these “tricky” fonts).

Structs

FontId 🔒

TableId 🔒

Constants

HINT_RELIANT_IDS 🔒

https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/truetype/ttobjs.c#L314

HINT_RELIANT_NAMES 🔒

https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/truetype/ttobjs.c#L180

MAX_HINT_RELIANT_NAME_LEN 🔒

Functions

is_hint_reliant_by_name 🔒

matches_hint_reliant_id_list 🔒

Checks for fonts that require hinting based on the length and checksum of the cvt, fpgm and prep tables.

matches_hint_reliant_name_list 🔒

Is this name on the list of fonts that require hinting?

require_interpreter 🔒

skip_pdf_random_tag 🔒

Fonts embedded in PDFs add random prefixes. Strip these for tricky font comparison purposes.