skrifa::outline

Module path

TrueType style outline to path conversion.

Structs

ContourPoint 🔒

Combination of point coordinates and flags.

Enums

PendingState 🔒

ToPathError

Errors that can occur when converting an outline to a path.

Functions

contour_to_path 🔒

Generates a path from an iterator of contour points.

contour_to_path_freetype 🔒

contour_to_path_harfbuzz 🔒

to_path 🔒

Converts a glyf outline described by points, flags and contour end points to a sequence of path elements and invokes the appropriate callback on the given pen for each.