skrifa::string

Struct StringId

pub struct StringId(/* private fields */);

Identifier for an informational string (or name).

A set of predefined identifiers exist for accessing names and other various metadata about the font and those are provided as associated constants on this type.

IDs 26 to 255, inclusive, are reserved for future standard names. IDs 256 to 32767, inclusive, are reserved for font-specific names such as those referenced by a font’s layout features.

For more detail, see https://learn.microsoft.com/en-us/typography/opentype/spec/name#name-ids

Implementations

impl NameId

pub const COPYRIGHT_NOTICE: NameId

Copyright notice.

pub const FAMILY_NAME: NameId

Font family name.

The font family name is used in combination with font subfamily name (ID 2), and should be shared among at most four fonts that differ only in weight or style (as described below).

pub const SUBFAMILY_NAME: NameId

Font subfamily name.

The font subfamily name is used in combination with font family name (ID 1), and distinguishes the fonts in a group with the same font family name. This should be used for style and weight variants only (as described below).

pub const UNIQUE_ID: NameId

Unique font identifier.

pub const FULL_NAME: NameId

Full font name that reflects all family and relevant subfamily descriptors.

The full font name is generally a combination of IDs 1 and 2, or of IDs 16 and 17, or a similar human-readable variant.

pub const VERSION_STRING: NameId

Version string.

Should begin with the syntax “Version number.number” (upper case, lower case, or mixed, with a space between “Version” and the number).

pub const POSTSCRIPT_NAME: NameId

PostScript name for the font.

ID 6 specifies a string which is used to invoke a PostScript language font that corresponds to this OpenType font. When translated to ASCII, the name string must be no longer than 63 characters and restricted to the printable ASCII subset, codes 33 to 126, except for the 10 characters ‘[’, ‘]’, ‘(’, ‘)’, ‘{’, ‘}’, ‘<’, ‘>’, ‘/’, ‘%’.

pub const TRADEMARK: NameId

Trademark; this is used to save any trademark notice/information for this font.

Such information should be based on legal advice. This is distinctly separate from the copyright.

pub const MANUFACTURER: NameId

Manufacturer name.

pub const DESIGNER: NameId

Name of the designer of the typeface.

pub const DESCRIPTION: NameId

Description of the typeface.

Can contain revision information, usage recommendations, history, features, etc.

pub const VENDOR_URL: NameId

URL of font vendor (with protocol, e.g., http://, ftp://).

If a unique serial number is embedded in the URL, it can be used to register the font.

pub const DESIGNER_URL: NameId

URL of typeface designer (with protocol, e.g., http://, ftp://).

pub const LICENSE_DESCRIPTION: NameId

License description.

A description of how the font may be legally used, or different example scenarios for licensed use. This field should be written in plain language, not legalese.

pub const LICENSE_URL: NameId

URL where additional licensing information can be found.

pub const TYPOGRAPHIC_FAMILY_NAME: NameId

Typographic family name.

The typographic family grouping doesn’t impose any constraints on the number of faces within it, in contrast with the 4-style family grouping (ID 1), which is present both for historical reasons and to express style linking groups.

pub const TYPOGRAPHIC_SUBFAMILY_NAME: NameId

Typographic subfamily name.

This allows font designers to specify a subfamily name within the typographic family grouping. This string must be unique within a particular typographic family.

pub const COMPATIBLE_FULL_NAME: NameId

Compatible full (Macintosh only).

On the Macintosh, the menu name is constructed using the FOND resource. This usually matches the full name. If you want the name of the font to appear differently than the Full Name, you can insert the compatible full name in ID 18.

pub const SAMPLE_TEXT: NameId

Sample text.

This can be the font name, or any other text that the designer thinks is the best sample to display the font in.

pub const POSTSCRIPT_CID_NAME: NameId

PostScript CID findfont name.

Its presence in a font means that the ID 6 holds a PostScript font name that is meant to be used with the “composefont” invocation in order to invoke the font in a PostScript interpreter.

pub const WWS_FAMILY_NAME: NameId

WWS family name.

Used to provide a WWS-conformant family name in case the entries for IDs 16 and 17 do not conform to the WWS model.

pub const WWS_SUBFAMILY_NAME: NameId

WWS subfamily name.

Used in conjunction with ID 21, this ID provides a WWS-conformant subfamily name (reflecting only weight, width and slope attributes) in case the entries for IDs 16 and 17 do not conform to the WWS model.

pub const LIGHT_BACKGROUND_PALETTE: NameId

Light background palette.

This ID, if used in the CPAL table’s Palette Labels Array, specifies that the corresponding color palette in the CPAL table is appropriate to use with the font when displaying it on a light background such as white.

pub const DARK_BACKGROUND_PALETTE: NameId

Dark background palette.

This ID, if used in the CPAL table’s Palette Labels Array, specifies that the corresponding color palette in the CPAL table is appropriate to use with the font when displaying it on a dark background such as black.

pub const VARIATIONS_POSTSCRIPT_NAME_PREFIX: NameId

Variations PostScript name prefix.

If present in a variable font, it may be used as the family prefix in the PostScript Name Generation for Variation Fonts algorithm.

pub const LAST_RESERVED_NAME_ID: NameId

The last value that is explicitly reserved for standard names.

Values after this are available to be used for font-specific names.

See https://learn.microsoft.com/en-us/typography/opentype/spec/name#name-ids

pub const LAST_ALLOWED_NAME_ID: NameId

The last value that is available for font-specific names.

See https://learn.microsoft.com/en-us/typography/opentype/spec/name#name-ids

impl NameId

pub const fn new(raw: u16) -> NameId

Create a new identifier from a raw u16 value.

pub fn predefined() -> impl Iterator<Item = NameId> + Clone

Returns an iterator over the set of predefined identifiers according to the specification.

pub const fn to_u16(self) -> u16

Return the identifier as a u16.

pub const fn is_reserved(self) -> bool

Returns true if self is in the range 0..=255 (reserved for standard names.)

pub const fn checked_add(self, rhs: u16) -> Option<NameId>

Returns a new NameId by adding rhs to the current value.

Returns None if the result would be greater than Self::LAST_ALLOWED_NAME_ID.

pub const fn to_be_bytes(self) -> [u8; 2]

Return the memory representation of this identifier as a byte array in big-endian (network) byte order.