skrifa

Struct AxisCollection

pub struct AxisCollection<'a> { /* private fields */ }

Collection of axes in a variable font.

Converts user (fvar) locations to normalized locations. See Self::location.

See the Axis type for more detail.

Implementations

impl<'a> AxisCollection<'a>

pub fn new(font: &FontRef<'a>) -> Self

Creates a new axis collection from the given font.

pub fn len(&self) -> usize

Returns the number of variation axes in the font.

pub fn is_empty(&self) -> bool

Returns true if the collection is empty.

pub fn get(&self, index: usize) -> Option<Axis>

Returns the axis at the given index.

pub fn get_by_tag(&self, tag: Tag) -> Option<Axis>

Returns the axis with the given tag.

Examples

let opsz = Tag::new(b"opsz");
assert_eq!(font.axes().get_by_tag(opsz).unwrap().tag(), opsz);

pub fn location<I>(&self, settings: I) -> Location
where I: IntoIterator, I::Item: Into<VariationSetting>,

Given an iterator of variation settings in user space, computes an ordered sequence of normalized coordinates.

Setting selectors that donโt match an axis are ignored.

Setting values are clamped to the range of their associated axis before normalization.

If more than one setting for an axis is provided, the last one is used.

Omitted settings are set to 0.0, representing the default position in variation space.

Examples

let location = font.axes().location([("wght", 250.0), ("wdth", 75.0)]);

pub fn location_to_slice<I>( &self, settings: I, location: &mut [NormalizedCoord], )
where I: IntoIterator, I::Item: Into<VariationSetting>,

Given an iterator of variation settings in user space, computes an ordered sequence of normalized coordinates and stores them in the target slice.

Setting selectors that donโt match an axis are ignored.

Setting values are clamped to the range of their associated axis before normalization.

If more than one setting for an axis is provided, the last one is used.

If no setting for an axis is provided, the associated coordinate is set to the normalized value 0.0, representing the default position in variation space.

Examples

let axes = font.axes();
let mut location = vec![NormalizedCoord::default(); axes.len()];
axes.location_to_slice([("wght", 250.0), ("wdth", 75.0)], &mut location);

pub fn filter<I>( &self, settings: I, ) -> impl Iterator<Item = VariationSetting> + Clone
where I: IntoIterator, I::Item: Into<VariationSetting>,

Given an iterator of variation settings in user space, returns a new iterator yielding those settings that are valid for this axis collection.

Setting selectors that donโt match an axis are dropped.

If more than one setting for an axis is provided, the last one is retained.

Setting values are clamped to the range of their associated axis.

Examples

// Assuming a font contains a single "wght" (weight) axis with range
// 100-900:
let axes = font.axes();
let filtered: Vec<_> = axes
    .filter([("wght", 400.0), ("opsz", 100.0), ("wght", 1200.0)])
    .collect();
// The first "wght" and "opsz" settings are dropped and the final
// "wght" axis is clamped to the maximum value of 900.
assert_eq!(&filtered, &[("wght", 900.0).into()]);

pub fn iter(&self) -> impl Iterator<Item = Axis> + 'a + Clone

Returns an iterator over the axes in the collection.