Module skrifa::charmap

Mapping of characters to nominal glyph identifiers.

The functionality in this module provides a 1-to-1 mapping from Unicode characters (or Unicode variation sequences) to nominal or “default” internal glyph identifiers for a given font. This is a necessary first step, but generally insufficient for proper layout of complex text or even simple text containing diacritics and ligatures.

Comprehensive mapping of characters to positioned glyphs requires a process called shaping. For more detail, see: Why do I need a shaping engine?

Structs

Charmap

Mapping of characters to nominal glyph identifiers.

MappingIndex

Cacheable indices of selected mapping tables for materializing a character map.

Mappings

Iterator over all mappings of character to nominal glyph identifier in a character map.

VariantMappings

Iterator over all mappings of character and variation selector to nominal glyph identifier in a character map.

Enums

MapVariant

Result of mapping a codepoint with a variation selector.