skrifa::outline::autohint

Module hint

Entry point to hinting algorithm.

Modules

edges 🔒

Edge hinting.

outline 🔒

Apply edge hints to an outline.

Structs

EdgeMetrics 🔒

Captures adjusted horizontal scale and outer edge positions to be used for horizontal metrics adjustments.

HintedMetrics 🔒

Functions

hint_outline 🔒

Applies the complete hinting process to a latin outline.