skrifa::outline::autohint

Module metrics

Autohinting specific metrics.

Modules

blues 🔒

Latin blue values.

scale 🔒

Metrics scaling.

widths 🔒

Latin standard stem width computation.

Structs

Scale 🔒

Captures scaling parameters which may be modified during metrics computation.

ScaledAxisMetrics 🔒

Scaled metrics for a single axis.

ScaledStyleMetrics 🔒

Scaled metrics for a single style and script.

ScaledWidth 🔒

UnscaledAxisMetrics 🔒

Unscaled metrics for a single axis.

UnscaledStyleMetrics 🔒

Unscaled metrics for a single style and script.

WidthMetrics 🔒

Enums

UnscaledStyleMetricsSet 🔒

The set of unscaled style metrics for a single font.

Constants

MAX_WIDTHS 🔒

Maximum number of widths, same for Latin and CJK.

Functions

fixed_div 🔒

fixed_mul 🔒

fixed_mul_div 🔒

pix_floor 🔒

pix_round 🔒

sort_and_quantize_widths 🔒

Type Aliases

ScaledWidths 🔒

UnscaledWidths 🔒