skrifa::outline::autohint::outline

Struct Outline

pub(crate) struct Outline {
    pub units_per_em: i32,
    pub orientation: Option<Orientation>,
    pub points: SmallVec<Point, MAX_INLINE_POINTS>,
    pub contours: SmallVec<Contour, MAX_INLINE_CONTOURS>,
    pub advance: i32,
}

Fields
units_per_em: i32orientation: Option<Orientation>points: SmallVec<Point, MAX_INLINE_POINTS>contours: SmallVec<Contour, MAX_INLINE_CONTOURS>advance: i32
Implementations

impl Outline

pub fn fill( &mut self, glyph: &OutlineGlyph<'_>, coords: &[F2Dot14], ) -> Result<(), DrawError>

Fills the outline from the given glyph.

pub fn scale(&mut self, scale: &Scale)

Applies dimension specific scaling factors and deltas to each point in the outline.

pub fn clear(&mut self)

pub fn to_path( &self, style: PathStyle, pen: &mut impl OutlinePen, ) -> Result<(), ToPathError>

impl Outline

fn link_points(&mut self)

Sets next and previous indices for each point.

fn mark_near_points(&mut self, near_limit: i32)

Computes the near flag for each contour.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/57617782464411201ce7bbc93b086c1b4d7d84a5/src/autofit/afhints.c#L1017

fn compute_directions(&mut self, near_limit: i32)

Compute directions of in and out vectors.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/57617782464411201ce7bbc93b086c1b4d7d84a5/src/autofit/afhints.c#L1064

fn simplify_topology(&mut self)

Simplify so that we can identify local extrema more reliably.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/57617782464411201ce7bbc93b086c1b4d7d84a5/src/autofit/afhints.c#L1181

fn check_remaining_weak_points(&mut self)

Check for remaining weak points.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/57617782464411201ce7bbc93b086c1b4d7d84a5/src/autofit/afhints.c#L1226

fn compute_orientation(&mut self)

Computes the overall winding order of the outline.

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/57617782464411201ce7bbc93b086c1b4d7d84a5/src/base/ftoutln.c#L1049