skrifa::outline::autohint::topo

Module segments

Segment computation and linking.

A segment is a series of at least two consecutive points that are appropriately aligned along a coordinate axis.

The linking stage associates pairs of segments to form stems and identifies serifs with a post-process pass.

Structs

State ğŸ”’

Capture current and previous state while computing segments.

Constants

MAX_SCORE ğŸ”’

MIN_SCORE ğŸ”’

Functions

_detect_round_segments_cjk ğŸ”’

Performs the additional step of detecting round segments for the CJK script group.

adjust_segment_heights ğŸ”’

Slightly increase the height of segments when it makes sense to better detect and ignore serifs.

assign_point_uvs ğŸ”’

Set the (u, v) values to font unit coords for each point depending on the axis dimension.

build_segments ğŸ”’

Build the set of segments for each contour.

compute_segments ğŸ”’

Computes segments for the Latin writing system.

link_segments ğŸ”’

Link segments to form stems and serifs.

link_segments_cjk ğŸ”’

Link segments to form stems and serifs for the CJK script group.

link_segments_default ğŸ”’

Link segments to form stems and serifs.