skrifa::outline::cff

Module hint

CFF hinting.

Structs

BlueZone 🔒

See https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/psaux/psblues.h#L129

Hint 🔒

https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/psaux/psblues.h#L118

HintMap 🔒

Collection of adjusted hint edges.

HintMask 🔒

Bitmask that specifies which hints are currently active.

HintParams 🔒

Parameters used to generate the stem and counter zones for the hinting algorithm.

HintState 🔒

Hinting state for a PostScript subfont.

HintingSink 🔒

StemHint 🔒

https://gitlab.freedesktop.org/freetype/freetype/-/blob/80a507a6b8e3d2906ad2c8ba69329bd2fb2a85ef/src/psaux/pshints.h#L85

Constants

EPSILON 🔒

GHOST_BOTTOM 🔒

GHOST_TOP 🔒

HINT_MASK_SIZE 🔒

ICF_BOTTOM 🔒

ICF_TOP 🔒

LOCKED 🔒

MAX_BLUES 🔒

MAX_BLUE_ZONES 🔒

MAX_HINTS 🔒

MAX_OTHER_BLUES 🔒

MIN_COUNTER 🔒

PAIR_BOTTOM 🔒

PAIR_TOP 🔒

SYNTHETIC 🔒

Functions

half 🔒

midpoint 🔒

Computes midpoint between a and b, avoiding overflow if the sum of the high 16 bits exceeds i16::MAX.

msb_mask 🔒

Returns a bit mask for the selected bit with the most significant bit at index 0.

trunc 🔒

FreeType converts from 16.16 to 26.6 by truncation. We keep our values in 16.16 so simply zero the low 10 bits to match the precision when converting to f32.

twice 🔒