skrifa::outline::glyf::deltas

Function composite_glyph

pub(super) fn composite_glyph<D: PointCoord>(
    gvar: &Gvar<'_>,
    glyph_id: GlyphId,
    coords: &[F2Dot14],
    deltas: &mut [Point<D>],
) -> Result<(), ReadError>

Compute a set of deltas for the component offsets of a composite glyph.

Interpolation is meaningless for component offsets so this is a specialized function that skips the expensive bits.