skrifa::outline::path

Function to_path

pub(crate) fn to_path<C: PointCoord>(
    points: &[Point<C>],
    flags: &[PointFlags],
    contours: &[u16],
    path_style: PathStyle,
    pen: &mut impl OutlinePen,
) -> Result<(), ToPathError>

Converts a glyf outline described by points, flags and contour end points to a sequence of path elements and invokes the appropriate callback on the given pen for each.

The input points can have any coordinate type that implements PointCoord. Output points are always generated in f32.

This is roughly equivalent to FT_Outline_Decompose.

See contour_to_path for a more general function that takes an iterator if your outline data is in a different format.